Improve smesh GUI display() speed for viewing mesh
Primarily, VTKViewer_GeometryFilter
is delegated  to  vtkGeometryFilter
used by ParaView.

At this point  typical  performance improvements are 5-10x over
the mesh display, this  dependes if  multithreaded  version  of
ParaView (consequently VTK) is used or not

Use old routines for nonlinear cells (quadratic 1D and 2D):
This should fix arc visulization
for quadratic elements

Bypass to old method for Polyhedron datat:
Multi-threding of VTK causes problems
for Polyhedron data, the code at  the
moment does not seem  stable.  To  be
investigated in future, if VTK can be
patched.

Addition to non-delegation list:
Add all 3d quadratic, biquadratic and
triquadratic elements to the list  of
elements  which  have  to  be handled
natively.